+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # Changelog
 
+## v1.0.3 rpetit3/sizemeup "Straightedge" 2024/09/30
+
+- allow querying of species name or taxid (replaces `--species` with `--query`)
+- Added `category` column to the output of `sizemeup`
+
 ## v1.0.2 rpetit3/sizemeup "Yardstick" 2024/09/29
 
 - fix issue where `--outdir` was used when writing files